+++ tests/scripts/script1.py
@@ -1,7 +1,6 @@
-#!/usr/bin/env click-odoo
+#!/usr/bin/env dodoo
 from __future__ import print_function
-
 
 env = env  # noqa
 
-print(env['res.users'].search([('login', '=', 'admin')]).login)
+print(env["res.users"].search([("login", "=", "admin")]).login)