+++ tests/scripts/script2.py
@@ -1,7 +1,7 @@
-#!/usr/bin/env click-odoo
+#!/usr/bin/env dodoo
 from __future__ import print_function
-import sys
 
+import sys
 
 env = env  # noqa